requirements.txt
azure-identity~=1.20.0
azure-storage-blob~=12.24.0
psutil~=7.0.0